rename module `executor` to `from_proto`
