+++ src/stream/src/from_proto/chain.rs
@@ -20,7 +20,7 @@ pub struct ChainExecutorBuilder;
 impl ExecutorBuilder for ChainExecutorBuilder {
     fn new_boxed_executor(
         mut params: ExecutorParams,
-        node: &stream_plan::StreamNode,
+        node: &StreamNode,
         _store: impl StateStore,
         stream: &mut LocalStreamManagerCore,
     ) -> Result<BoxedExecutor> {
