+++ src/stream/src/from_proto/filter.rs
@@ -12,23 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::error::Result;
-use risingwave_common::try_match_expand;
 use risingwave_expr::expr::build_from_prost;
-use risingwave_pb::stream_plan;
-use risingwave_pb::stream_plan::stream_node::NodeBody;
-use risingwave_storage::StateStore;
 
-use crate::executor::ExecutorBuilder;
-use crate::executor_v2::{BoxedExecutor, Executor, FilterExecutor};
-use crate::task::{ExecutorParams, LocalStreamManagerCore};
+use super::*;
+use crate::executor_v2::FilterExecutor;
 
 pub struct FilterExecutorBuilder;
 
 impl ExecutorBuilder for FilterExecutorBuilder {
     fn new_boxed_executor(
         mut params: ExecutorParams,
-        node: &stream_plan::StreamNode,
+        node: &StreamNode,
         _store: impl StateStore,
         _stream: &mut LocalStreamManagerCore,
     ) -> Result<BoxedExecutor> {
