+++ src/stream/src/from_proto/hash_agg.rs
@@ -57,7 +57,7 @@ pub struct HashAggExecutorBuilder;
 impl ExecutorBuilder for HashAggExecutorBuilder {
     fn new_boxed_executor(
         mut params: ExecutorParams,
-        node: &stream_plan::StreamNode,
+        node: &StreamNode,
         store: impl StateStore,
         _stream: &mut LocalStreamManagerCore,
     ) -> Result<BoxedExecutor> {
