+++ src/stream/src/from_proto/hash_join.rs
@@ -28,7 +28,7 @@ pub struct HashJoinExecutorBuilder;
 impl ExecutorBuilder for HashJoinExecutorBuilder {
     fn new_boxed_executor(
         mut params: ExecutorParams,
-        node: &stream_plan::StreamNode,
+        node: &StreamNode,
         store: impl StateStore,
         _stream: &mut LocalStreamManagerCore,
     ) -> Result<BoxedExecutor> {
