+++ src/stream/src/from_proto/local_simple_agg.rs
@@ -1,0 +1,44 @@
+// Copyright 2022 Singularity Data
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use super::*;
+use crate::executor_v2::aggregation::AggCall;
+use crate::executor_v2::LocalSimpleAggExecutor;
+use crate::task::build_agg_call_from_prost;
+
+pub struct LocalSimpleAggExecutorBuilder;
+
+impl ExecutorBuilder for LocalSimpleAggExecutorBuilder {
+    fn new_boxed_executor(
+        mut params: ExecutorParams,
+        node: &StreamNode,
+        _store: impl StateStore,
+        _stream: &mut LocalStreamManagerCore,
+    ) -> Result<BoxedExecutor> {
+        let node = try_match_expand!(node.get_node_body().unwrap(), NodeBody::LocalSimpleAgg)?;
+        let agg_calls: Vec<AggCall> = node
+            .get_agg_calls()
+            .iter()
+            .map(build_agg_call_from_prost)
+            .try_collect()?;
+
+        Ok(LocalSimpleAggExecutor::new(
+            params.input.remove(0),
+            agg_calls,
+            params.pk_indices,
+            params.executor_id,
+        )?
+        .boxed())
+    }
+}
