+++ src/stream/src/from_proto/lookup.rs
@@ -24,7 +24,7 @@ pub struct LookupExecutorBuilder;
 impl ExecutorBuilder for LookupExecutorBuilder {
     fn new_boxed_executor(
         mut params: ExecutorParams,
-        node: &stream_plan::StreamNode,
+        node: &StreamNode,
         store: impl StateStore,
         _stream: &mut LocalStreamManagerCore,
     ) -> Result<BoxedExecutor> {
