+++ src/stream/src/from_proto/mview.rs
@@ -23,7 +23,7 @@ pub struct MaterializeExecutorBuilder;
 impl ExecutorBuilder for MaterializeExecutorBuilder {
     fn new_boxed_executor(
         mut params: ExecutorParams,
-        node: &stream_plan::StreamNode,
+        node: &StreamNode,
         store: impl StateStore,
         _stream: &mut LocalStreamManagerCore,
     ) -> Result<BoxedExecutor> {
@@ -60,7 +60,7 @@ pub struct ArrangeExecutorBuilder;
 impl ExecutorBuilder for ArrangeExecutorBuilder {
     fn new_boxed_executor(
         mut params: ExecutorParams,
-        node: &stream_plan::StreamNode,
+        node: &StreamNode,
         store: impl StateStore,
         _stream: &mut LocalStreamManagerCore,
     ) -> Result<BoxedExecutor> {
