+++ src/stream/src/from_proto/top_n_appendonly.rs
@@ -12,23 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::error::Result;
-use risingwave_common::try_match_expand;
 use risingwave_common::util::sort_util::OrderPair;
-use risingwave_pb::stream_plan;
-use risingwave_pb::stream_plan::stream_node::NodeBody;
-use risingwave_storage::{Keyspace, StateStore};
 
-use crate::executor::ExecutorBuilder;
-use crate::executor_v2::{AppendOnlyTopNExecutor, BoxedExecutor, Executor};
-use crate::task::{ExecutorParams, LocalStreamManagerCore};
+use super::*;
+use crate::executor_v2::AppendOnlyTopNExecutor;
 
 pub struct AppendOnlyTopNExecutorBuilder;
 
 impl ExecutorBuilder for AppendOnlyTopNExecutorBuilder {
     fn new_boxed_executor(
         mut params: ExecutorParams,
-        node: &stream_plan::StreamNode,
+        node: &StreamNode,
         store: impl StateStore,
         _stream: &mut LocalStreamManagerCore,
     ) -> Result<BoxedExecutor> {
