+++ src/stream/src/lib.rs
@@ -44,6 +44,6 @@
 extern crate log;
 
 pub mod common;
-pub mod executor;
 pub mod executor_v2;
+pub mod from_proto;
 pub mod task;
