+++ src/stream/src/task/stream_manager.rs
@@ -35,13 +35,13 @@ use tokio::sync::oneshot;
 use tokio::task::JoinHandle;
 
 use super::{unique_executor_id, unique_operator_id, CollectResult, ComputeClientPool};
-use crate::executor::create_executor;
 use crate::executor_v2::aggregation::{AggArgs, AggCall};
 use crate::executor_v2::dispatch::*;
 use crate::executor_v2::merge::RemoteInput;
 use crate::executor_v2::monitor::StreamingMetrics;
 use crate::executor_v2::receiver::ReceiverExecutor;
 use crate::executor_v2::*;
+use crate::from_proto::create_executor;
 use crate::task::{
     ActorId, ConsumableChannelPair, SharedContext, StreamEnvironment, UpDownActorIds,
     LOCAL_OUTPUT_CHANNEL_SIZE,
